+++ Assets/Scripts/TradeItem.cs
@@ -2,16 +2,16 @@
 
 public class TradeItem
 {
-	public TradeItem(Item item, decimal amount, decimal pricePerItem, PlayerController player)
-	{
-		Item = item;
-		Amount = amount;
-		PricePerItem = pricePerItem;
+    public TradeItem(Item item, double amount, double pricePerItem, PlayerController player)
+    {
+        Item = item;
+        Amount = amount;
+        PricePerItem = pricePerItem;
         Player = player;
-	}
+    }
 
-	public Item Item { get; }
-	public decimal Amount { get; }
-	public decimal PricePerItem { get; }
-	public PlayerController Player { get; }
+    public Item Item { get; }
+    public double Amount { get; }
+    public double PricePerItem { get; }
+    public PlayerController Player { get; }
 }
